# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DDD asosida qurilgan e-commerce marketplace backend",
        "title": "Uzum Clone API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/orders": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin tomonidan cart'siz, to'g'ridan-to'g'ri buyurtma yaratadi (masalan, telefon orqali qabul qilingan yoki do'kondagi offline savdo). Catalog'dan joriy nom/narxni \"surat\" qiladi va zaxirani atomik ravishda kamaytiradi (Checkout bilan bir xil mantiq). Faqat admin uchun.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Qo'lda buyurtma yaratish (offline savdo)",
                "parameters": [
                    {
                        "description": "Mijoz ma'lumotlari, mahsulotlar ro'yxati va boshlang'ich to'lov holati",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateManualOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Buyurtma yaratildi",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.OrderResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Noto'g'ri so'rov, bo'sh ro'yxat yoki noto'g'ri holat/manzil/telefon qiymati",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Autentifikatsiya talab qilinadi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Faqat admin uchun",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Mahsulot topilmadi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "409": {
                        "description": "Mahsulot uchun yetarli zaxira yo'q",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Ichki server xatosi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/admin/orders/{id}/cancel": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Buyurtmani bekor qiladi va Checkout/CreateManualOrder paytida kamaytirilgan zaxirani Catalog'ga qaytaradi. Faqat Preparing holatidagi buyurtmalar bekor qilinishi mumkin - courier'ga topshirilgan yoki yetkazilgan buyurtmani bu yo'l bilan bekor qilib bo'lmaydi. Faqat admin uchun.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Buyurtmani bekor qilish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Buyurtma ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bekor qilingan buyurtma",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.OrderResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Autentifikatsiya talab qilinadi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Faqat admin uchun",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Buyurtma topilmadi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "409": {
                        "description": "Courier'ga topshirilgan yoki yetkazilgan buyurtmani bekor qilib bo'lmaydi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Ichki server xatosi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
        "/checkout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Joriy foydalanuvchining savatidagi mahsulotlarni buyurtmaga aylantiradi: Catalog'dan joriy nom/narxni \"surat\" qiladi, zaxirani atomik ravishda kamaytiradi va savatni bo'shatadi",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Savatni buyurtmaga aylantirish",
                "parameters": [
                    {
                        "description": "Yetkazib berish ma'lumotlari",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CheckoutRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Buyurtma yaratildi",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.OrderResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Noto'g'ri so'rov, bo'sh savat yoki yaroqsiz telefon/manzil",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Autentifikatsiya talab qilinadi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Mahsulot topilmadi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "409": {
                        "description": "Mahsulot uchun yetarli zaxira yo'q",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Ichki server xatosi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
        "/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Joriy autentifikatsiya qilingan foydalanuvchining barcha buyurtmalari ro'yxati (buyurtma holati sahifasi uchun)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Mening buyurtmalarim",
                "responses": {
                    "200": {
                        "description": "Buyurtmalar ro'yxati",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/http.OrderResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Autentifikatsiya talab qilinadi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Ichki server xatosi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/orders/admin": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Barcha foydalanuvchilarning barcha buyurtmalari ro'yxati. Faqat admin uchun.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Barcha buyurtmalar (admin navbati)",
                "responses": {
                    "200": {
                        "description": "Buyurtmalar ro'yxati",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/http.OrderResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Autentifikatsiya talab qilinadi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Faqat admin uchun",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Ichki server xatosi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Bitta buyurtmani ID bo'yicha oladi. Faqat buyurtma egasi yoki admin ko'ra oladi.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Buyurtma tafsilotlari",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Buyurtma ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Buyurtma",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.OrderResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Autentifikatsiya talab qilinadi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Bu buyurtmani ko'rish huquqi yo'q",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Buyurtma topilmadi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Ichki server xatosi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.CheckoutRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "http.CreateManualOrderRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.ManualOrderItemRequest"
                    }
                },
                "note": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "http.ManualOrderItemRequest": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "http.OrderItemResponse": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "unit_price": {
                    "type": "number"
                }
            }
        },
        "http.OrderResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivery_status": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.OrderItemResponse"
                    }
                },
                "note": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "total_currency": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}